Fix overriding a rule with a disable one
If we want to disable a rule from an inherited set of rules through
overriding it in the child ruleset the disabled rule is not created but
the rule from the parent is left on the set of active rules.

This fix this situation by deleting the rule (if exists) from the set of
active rules if the most specific rule is declared as disabled.