Make 'verbose' explicit in lint test to avoid failures
The tests in lint.test.js fails if you have a speccy.yaml config file
with 'verbose: true' because all the tests spying on the messages
sent to the console through log(), warn() and error() methods and
asserting the quantity of calls.

Making it explicit as the default config for those tests allows us to
run them even if we have a speccy.yaml in the root directory.