Config: prevent meaningless empty arrays override options in file
It fixes the impossibility that the 'rules' and 'skip' options specified
in speccy.yaml config file are used when these options are not
override from command line.

The Commander.js configuration has an empty array as a default/initial
value for 'rules' and 'skip'. If you execute the lint command from the
command line without specifying those arguments, when the config is
loaded those 'supplied' literal config values (empty arrays) overrides
the ones defined in a speccy.yaml file (because they are not 'undefined').

There is a test validating the config hierarchy but, as the configuration
passed to init() is handcrafted, it wasn't simulating the output of
Commander.js read options.
Now, a test simulating this case was added.